% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2.R
\name{co2_ch4_monthly}
\alias{co2_ch4_monthly}
\title{Fetch monthly global CH4 (methane) data}
\usage{
co2_ch4_monthly()
}
\value{
tibble: year, month, decimal, average, average_unc, trend, trend_unc
}
\description{
Fetch monthly global CH4 (methane) data
}
