% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2.R
\name{co2_global_trend}
\alias{co2_global_trend}
\title{Fetch daily global CO2 trend data}
\usage{
co2_global_trend()
}
\value{
tibble: date (Date), smoothed (numeric ppm), trend (numeric ppm)
}
\description{
Smoothed and trend values from NOAA GML. Updated daily.
}
